maxStoreSize is 64bit and can have string param in config that support kb,mb,gb suffixes

+++ src/system/scheduler.hpp
@@ -88,7 +88,7 @@ protected:
   bool running;
   bool loadup;
   bool rolling;
-  int maxStoreSize;
+  uint64_t maxStoreSize;
   time_t minRollTime,lastRollTime;
   IdSeqPairList snap;
   std::string rolFile;
